Restructured app structure, implemented auth and users
Restructured the app to contain db_modules, with a single instance of the db firestore client that is attached to app.config['db']. This client object is then shared across all blueprints, and passed to database methods located in db_modules.

I began implementing auth services for sign up and login, which establishes a user session. Right now we can't completely implement or test because we don't have any credentials setup with Apple Log In API

Currently can't run the application because I'm getting a file not found error when the app looks for firestore_sa.json